Test/json boards

+++ src/lib/actions.ts
@@ -11,7 +11,6 @@ import {
 	ResetPasswordFormSchema,
 	EditProfileFormSchema,
 	DeleteAccountFormSchema,
-    UploadSchema,
     RenameBoardSchema,
     BoardFormState,
 } from "@/lib/definitions";
@@ -21,8 +20,8 @@ export async function signup(formState: UserFormState, formData: FormData) {
     const validatedFields = SignupFormSchema.safeParse({
 		displayName: formData.get("displayName"),
 		email: formData.get("email"),
-		password: formData.get("password"),
-		confirm: formData.get("confirm"),
+		newPassword: formData.get("newPassword"),
+		confirmPassword: formData.get("confirmPassword"),
 	});
 
 	if (!validatedFields.success) {
@@ -248,7 +247,10 @@ export async function updatePassword(formState: UserFormState, formData: FormDat
             },
         );
 
-        if (passwordError) throw passwordError;
+        if (passwordError) {
+            console.log(passwordError);
+            throw passwordError;
+        }
 
         revalidatePath("/");
 
@@ -287,7 +289,10 @@ export async function updateEmail(formState: UserFormState, formData: FormData) 
             .select("email")
             .eq("email", validatedFields.data.email);
 
-        if (emailExistsError) throw emailExistsError;
+        if (emailExistsError) {
+            console.log(emailExistsError)
+            throw emailExistsError;
+        }
 
         if (emailExistsData.length > 0) {
             return {
@@ -304,7 +309,7 @@ export async function updateEmail(formState: UserFormState, formData: FormData) 
             },
         });
 
-        if (updateError) throw updateError
+        if (updateError) throw updateError;
 
         revalidatePath("/");
 
@@ -413,27 +418,12 @@ export async function updateUserProfile(
     }
 }
 
-export async function uploadAvatar(formState: UserFormState, formData: FormData) {
-    const validatedFields = UploadSchema.safeParse({
-        avatar: formData.get("avatar"),
-    })
-
-    if (!validatedFields.success) {
-        return {
-            errors: validatedFields.error.flatten().fieldErrors,
-        }
-    }
-
+export async function uploadAvatar(userId: string, file: File) {
     try {
         const supabase = await createClient();
 
-        const { data: userData, error: userError } = await supabase.auth.getUser();
-
-        if (userError) throw userError;
-
-        const file = validatedFields.data.avatar;
         const fileExt = file.name.split(".").pop();
-        const filePath = `${userData.user.id}/avatar_${Date.now()}.${fileExt}`;
+        const filePath = `${userId}/avatar_${Date.now()}.${fileExt}`;
 
         const { error: uploadError } = await supabase.storage
             .from("avatars")
@@ -446,12 +436,57 @@ export async function uploadAvatar(formState: UserFormState, formData: FormData)
             .update({
                 avatar_path: filePath,
             })
-            .eq("id", userData.user.id);
+            .eq("id", userId);
 
         if (profileError) throw profileError;
 
         const { data: publicUrl } = await supabase.storage
             .from("avatars")
+            .getPublicUrl(filePath);
+
+        revalidatePath("/settings");
+
+        return {
+            publicUrl: publicUrl.publicUrl,
+        }
+
+    } catch {
+        return {
+            errorMessage: "We could not update your avatar. Please try again.",
+        };
+    }
+}
+
+export async function uploadCover(boardId: string, file: File) {
+    try {
+        const supabase = await createClient();
+
+        const fileExt = file.name.split(".").pop();
+        const filePath = `${boardId}/cover_${Date.now()}.${fileExt}`;
+
+        const { error: uploadError } = await supabase.storage
+            .from("covers")
+            .upload(filePath, file);
+
+        if (uploadError) {
+            console.log(uploadError.message)
+            throw uploadError;
+        }
+
+        const { error: boardError } = await supabase
+            .from("boards")
+            .update({
+                cover_path: filePath,
+            })
+            .eq("id", boardId);
+
+        if (boardError) {
+            console.log(boardError.message)
+            throw boardError;
+        }
+
+        const { data: publicUrl } = await supabase.storage
+            .from("covers")
             .getPublicUrl(filePath);
 
         revalidatePath("/");
@@ -462,7 +497,7 @@ export async function uploadAvatar(formState: UserFormState, formData: FormData)
 
     } catch {
         return {
-            errorMessage: "We could not update your avatar. Please try again.",
+            errorMessage: "We could not update this board's cover. Please try again.",
         };
     }
 }
@@ -500,14 +535,9 @@ export async function deleteAccount(formState: UserFormState, formData: FormData
             if (removeError) throw removeError;
         }
 
-        const { error: socialsDeleteError } = await supabase
-            .from("socials")
-            .delete()
-            .eq("profile_id", userData.user.id);
-
-        if (socialsDeleteError) throw socialsDeleteError;
-
-        const { error: deleteError } = await supabase.rpc("handle_delete_user");
+        const { error: deleteError } = await supabase.rpc(
+            "handle_delete_user"
+        );
 
         if (deleteError) throw deleteError;
 
@@ -532,26 +562,41 @@ export async function createBoard() {
 
     if (userError) throw userError;
 
-    const board_id = crypto.randomUUID();
+    const newBoardId = crypto.randomUUID();
 
     const { error: boardError } = await supabase.from("boards").insert({
-        board_id: board_id,
+        id: newBoardId,
         profile_id: userData.user.id,
-        title: "Untitled Board",
-    });
+    })
 
     if (boardError) throw boardError;
 
-    redirect("/board/" + board_id);
+    revalidatePath("/")
+    redirect("/board/" + newBoardId);
 }
 
 export async function deleteBoard(boardId: string) {
     const supabase = await createClient();
 
+    const { data: folderData, error: folderError } = await supabase.storage
+        .from("covers")
+        .list(boardId);
+
+    if (folderError) throw folderError;
+
+    if (folderData.length > 0) {
+        const files = folderData.map((file) => `${boardId}/${file.name}`);
+        const { error: removeError } = await supabase.storage
+            .from("covers")
+            .remove(files);
+
+        if (removeError) throw removeError;
+    }
+
     const { error: deleteError } = await supabase
         .from("boards")
         .delete()
-        .eq("board_id", boardId);
+        .eq("id", boardId);
 
     if (deleteError) throw deleteError;
 
@@ -564,7 +609,7 @@ export async function bookmarkBoard(boardId: string, currentlyBookmarked: boolea
     const { error: bookmarkError } = await supabase
         .from("boards")
         .update({bookmarked: !currentlyBookmarked})
-        .eq("board_id", boardId);
+        .eq("id", boardId);
 
     if (bookmarkError) throw bookmarkError;
 
@@ -588,7 +633,7 @@ export async function renameBoard(formState: BoardFormState, formData: FormData)
         const { error: renameError } = await supabase
             .from("boards")
             .update({title: validatedFields.data.title})
-            .eq("board_id", formState?.boardId!);
+            .eq("id", formState?.boardId!);
 
         if (renameError) throw renameError;
